requirements.txt
roboflow
supervision==0.22.0
google-generativeai
pillow
numpy
tqdm